+++ app/models/call_for_evidence_response_form_data.rb
@@ -1,16 +1,20 @@
 class CallForEvidenceResponseFormData < ApplicationRecord
+  include AssetData
+
   mount_uploader :file, ResponseDocumentUploader, mount_on: :carrierwave_file
 
   has_one :call_for_evidence_response_form
 
-  has_many :assets,
-           as: :assetable,
-           inverse_of: :assetable
-
   validates :file, presence: true
 
-  def auth_bypass_ids
-    [call_for_evidence_response_form.call_for_evidence_participation.call_for_evidence.auth_bypass_id]
+  delegate :auth_bypass_id, to: :attachable
+
+  def attachable
+    call_for_evidence_response_form || Attachable::Null.new
+  end
+
+  def attachments
+    [call_for_evidence_response_form || Attachment::Null.new]
   end
 
   def all_asset_variants_uploaded?
