+++ app/models/consultation_response_form_data.rb
@@ -1,16 +1,21 @@
 class ConsultationResponseFormData < ApplicationRecord
+  include AssetData
+
   mount_uploader :file, ResponseDocumentUploader, mount_on: :carrierwave_file
 
   has_one :consultation_response_form
-  has_many :assets,
-           as: :assetable,
-           inverse_of: :assetable
 
   validates :file, presence: true
 
-  def auth_bypass_ids
-    [consultation_response_form.consultation_participation.consultation.auth_bypass_id]
+  def attachable
+    consultation_response_form || Attachable::Null.new
   end
+
+  def attachments
+    [consultation_response_form || Attachment::Null.new]
+  end
+
+  delegate :auth_bypass_id, to: :attachable
 
   def all_asset_variants_uploaded?
     asset_variants = assets.map(&:variant).map(&:to_sym)
