+++ app/models/image_data.rb
@@ -10,13 +10,11 @@ class ImageData < ApplicationRecord
 
   include Replaceable
   include ImageKind
+  include AssetData
 
   SVG_CONTENT_TYPE = "image/svg+xml".freeze
 
   has_many :images
-  has_many :assets,
-           as: :assetable,
-           inverse_of: :assetable
 
   mount_uploader :file, ImageUploader, mount_on: :carrierwave_image
 
@@ -32,6 +30,30 @@ class ImageData < ApplicationRecord
   after_validation :prepend_filename_to_errors
 
   delegate :url, :content_type, to: :file
+
+  def deleted?
+    return true if attachable.blank?
+
+    attachable.is_a?(Attachable::Null) || attachable.deleted?
+  end
+
+  def needs_publishing?
+    return true if deleted?
+
+    attachments.size == 1 && attachments.first.attachable.publicly_visible?
+  end
+
+  def attachable
+    return Attachable::Null.new if images.empty?
+
+    images.last.edition
+  end
+
+  def attachments
+    return [Image.new] if images.empty?
+
+    images
+  end
 
   def filename
     file&.file&.filename
