+++ test/unit/app/models/consultation_response_form_data_test.rb
@@ -64,4 +64,18 @@ class ConsultationResponseFormDataTest < ActiveSupport::TestCase
 
     assert_not consultation_response_form_data.all_asset_variants_uploaded?
   end
+
+  test ".attachable returns associated CallForEvidence if it is exists" do
+    consultation = build(:consultation)
+    consultation_participation = build(:consultation_participation, consultation:)
+    consultation_response_form = build(:consultation_response_form, consultation_participation:)
+
+    assert consultation_response_form.attachable, consultation
+  end
+
+  test ".attachable returns Attachable::Null if no associated CallForEvidence" do
+    consultation_response_form = build(:consultation_response_form)
+
+    assert consultation_response_form.attachable, Attachable::Null
+  end
 end
